pyproject.toml
[project]
name = "shonen-cast-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.12",
    "groq>=0.37.1",
    "python-dotenv>=1.2.1",
    "assemblyai>=0.33.0",
    "sqlalchemy>=2.0.45",
    "tavily-python>=0.7.15",
    "fastapi>=0.124.4",
    "uvicorn>=0.38.0",
]
